Revert "Merge branch 'release_candidate'"
This reverts commit bef51aed032c0aaa5cfd80445bc4cf0d85b408b5, reversing
changes made to cf2772fab0af5573da775e7437e6acdca424f26e.

+++ extensions-builtin/Lora/network_ia3.py
@@ -17,7 +17,7 @@ class NetworkModuleIa3(network.NetworkModule):
         self.on_input = weights.w["on_input"].item()
 
     def calc_updown(self, orig_weight):
-        w = self.w.to(orig_weight.device)
+        w = self.w.to(orig_weight.device, dtype=orig_weight.dtype)
 
         output_shape = [w.size(0), orig_weight.size(1)]
         if self.on_input:
